+++ Functions/Context.ps1
@@ -80,7 +80,11 @@ function ContextImpl
     {
         Add-SetupAndTeardown -ScriptBlock $Fixture
         Invoke-TestGroupSetupBlocks -Scope $pester.Scope
-        $null = & $Fixture
+
+        do
+        {
+            $null = & $Fixture
+        } until ($true)
     }
     catch
     {