+++ Functions/SetupTeardown.Tests.ps1
@@ -156,3 +156,31 @@ Describe 'Finishing TestGroup Setup and Teardown tests' {
         $script:ContextAfterAllCounter   | Should Be 1
     }
 }
+
+
+if ($PSVersionTable.PSVersion.Major -ge 3)
+{
+    $thisTestScriptFilePath = $ExecutionContext.SessionState.Path.GetUnresolvedProviderPathFromPSPath($PSCommandPath)
+
+    Describe 'Script Blocks and file association (testing automatic variables)' {
+        BeforeEach {
+            $commandPath = $PSCommandPath
+        }
+
+        $beforeEachBlock = InModuleScope Pester {
+            $pester.BeforeEach[0].ScriptBlock
+        }
+
+        It 'Creates script block objects associated with the proper file' {
+            $scriptBlockFilePath = $ExecutionContext.SessionState.Path.GetUnresolvedProviderPathFromPSPath($beforeEachBlock.File)
+
+            $scriptBlockFilePath | Should Be $thisTestScriptFilePath
+        }
+
+        It 'Has the correct automatic variable values inside the BeforeEach block' {
+            $commandPath | Should Be $PSCommandPath
+        }
+    }
+}
+
+#Testing if failing setup or teardown will fail 'It' is done in the TestsRunningInCleanRunspace.Tests.ps1 file