+++ CHANGELOG.md
@@ -1,4 +1,21 @@
 ## Unreleased
+  - Fixed Describe's handling of TestName filter when multiple strings are passed to Invoke-Pester's -TestName parameter.
+  - Failing BeforeEach or AfterEach will fail the test [GH-326]
+  - Added BeOfType operator to the Should command. [GH-327]
+  - Fixed BeforeEach / etc parsing in PSv3+ so breakpoints and automatic variables ($PSCommandPath, etc) will work properly.  [GH-333]
+  - Fixed bug in 'Should Be' when comparing strings, and null or empty strings are piped in to the Should command.  [GH-333]
+  - Added some calls to Write-Progress in the It command.  [GH-322]
+  - Bug fix when mocking functions that are in the global scope; the original functions were being lost when the Describe block ended.  [GH-323]
+
+## 3.3.8 (April 15, 2015)
+  - Further mocking fixes around the use of $ExecutionContext in client scope.  [GH-307]
+
+## 3.3.7 (April 15, 2015)
+  - Added workaround for GetDynamicParameters() bug that was affecting mocks on the ActiveDirectory module in Windows 7. [GH-295]
+  - Revised Mocking code to avoid potential bugs when functions define parameters named $ExecutionContext or $MyInvocation. [GH-304]
+  - Mocked functions no longer call Get-MockDynamicParameters if the original function had no dynamicparam block. [GH-306]
+
+## 3.3.6 (March 19, 2015)
   - Fix for mocking aliases for commands that are in scopes that Pester can't normally see. [GH-267]
   - Added line information to test failure output in Should assertion failures. [GH-266]
   - Added support for passing named parameters or positional arguments to test scripts, and for calling test scripts that are not named *.Tests.ps1.  [GH-272]
@@ -6,6 +23,7 @@
   - Improved error message when InModuleScope finds multiple modules loaded with the same name. [GH-276]
   - Updated build script to allow for custom root folder in the nupkg. [GH-254]
   - Improved error messages for InModuleScope and Mock -ModuleName when multiple modules with the same name are loaded. Also enabled these commands to work if only one of the loaded modules is a Script module. [GH-278]
+  - Added some graceful handling of test code that has a misplaced break or continue statement. [GH-290]
 
 ## 3.3.5 (January 23, 2015)
   - Updated tests to allow PRs to be automatically tested, with status updates to GitHub, by our CI server.